fhir-jembi: reorganise tests
